pyproject.toml
[project]
name = "customer-support-rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb==1.1.0",
    "datasets==4.1.0",
    "fastapi==0.116.2",
    "ipykernel==6.30.1",
    "langchain==0.3.27",
    "langchain-community==0.3.29",
    "langchain-core==0.3.76",
    "langchain-google-genai==2.1.12",
    "langchain-pinecone==0.2.12",
    "langgraph==0.6.7",
    "pinecone==7.3.0",
    "pinecone-client==6.0.0",
    "python-dotenv==1.1.1",
    "ragas==0.3.4",
    "sentence-transformers==5.1.0",
    "tqdm==4.67.1",
    "transformers==4.56.1",
]
